Add check that destination URL exists
Until now, `seePermanentRedirectTo()` checked if A sent the `Location` header and a 301 status to redirect to B, but didn't check if B actually exists.

By default, this now happens, to see if B responds with a 200 status code, and that the URL hasn't changed (didn't silently redirect to C).

Fixes #8