Change how URL exists is checked
Whereas previously it would look through the PhpBrowser client history for the recent module, this now seems to be empty when using a Redirects Cest. Until I figure out how if there's a way to fix that, we simply look for the non-existence of the Location header in the URL. Since we also state to follow redirects, this is pretty much default behaviour anyway.